Improve the animations and scrolling
The first issue this commit resolves is the lack of animations in Safari
and iOS (and maybe some other browsers). This was due to the fact that
the component was not marked with `display: box`. I've also added a
polyfill for those browsers

Those browsers are also guilty of not implementing smooth scrolling on
the page, so I've added a polyfill for that as well.

Lastly, this commit also tweaks the animations a bit; they look much
nicer now.

Signed-off-by: Jord Nijhuis <jord@nijhuis.me>