Improve the checklist items for processing grounds
Up until now, we grouped all the processing grounds together, and
separated the additional requirements for the "permission" ground. This
made little sense, as you would get a message that you had valid grounds
for processing, even though you did not comply with all the extra
requirements.

Thus, all the items related to processing grounds are now grouped
together into one item that shows as a fail as soon as you miss one
additional requirement. The description of the item tells the user why
he has passed/failed the requirements.

Signed-off-by: Jord Nijhuis <jord@nijhuis.me>